+++ src/crud.js
@@ -1,3 +1,6 @@
+import { taskNotComplete, taskCompleted } from './status.js';
+
+const ul = document.querySelector('ul');
 const taskDescription = document.querySelector('#txt-input');
 const deleteAllBtn = document.querySelector('.clear-all');
 
@@ -8,6 +11,74 @@ const getFromLocalStorage = () => {
 
 const tasks = getFromLocalStorage();
 
+const createUi = (task) => {
+ 
+    const li = document.createElement('li');
+    li.setAttribute('id', task.index);
+    const lDiv = document.createElement('div');
+    lDiv.classList.add('ldiv');
+    const lDivSpan = document.createElement('span');
+    lDivSpan.classList.add('ldiv-spn');
+    const rDiv = document.createElement('div');
+    rDiv.classList.add('rdiv');
+    const pTask = document.createElement('span');
+    const checkbox = document.createElement('input');
+    checkbox.type = 'checkbox';
+    checkbox.name = 'task-box';
+    checkbox.value = task;
+    checkbox.classList.add('chkbx');
+
+    li.appendChild(lDiv);
+
+    lDiv.appendChild(lDivSpan);
+    lDiv.appendChild(pTask);
+    lDivSpan.appendChild(checkbox);
+
+    li.appendChild(rDiv);
+
+    rDiv.innerHTML = '<i class="fas fa-bars"></i>';
+
+    ul.appendChild(li);
+
+    li.addEventListener('mouseover', () => {
+      rDiv.innerHTML = '<i class="fas fa-trash-alt"></i>';
+    });
+
+    li.addEventListener('mouseout', () => {
+      rDiv.innerHTML = '<i class="fas fa-bars"></i>';
+    });
+
+    pTask.addEventListener('click', () => {
+      editTasks(pTask, task);
+    });
+
+    rDiv.addEventListener('click', () => {
+      deleteTask(rDiv.parentElement.id);
+    });
+
+    if (task.completed === true) {
+      pTask.innerText = task.description;
+      pTask.classList.add('cancel');
+      pTask.previousSibling.children[0].checked = true;
+    } else {
+      pTask.innerText = task.description;
+    }
+
+    checkbox.addEventListener('change', (e) => {
+      if (e.target.checked) {
+        taskCompleted(task);
+        e.target.parentElement.nextSibling.classList.add('cancel');
+        addToLocalStorage();
+      } else {
+        taskNotComplete(task);
+        e.target.parentElement.nextSibling.classList.remove('cancel');
+        addToLocalStorage();
+      }
+    });
+
+};
+
+
 const addToLocalStorage = () => {
   const storage = JSON.stringify(tasks);
   localStorage.setItem('lstore', storage);
@@ -15,13 +86,14 @@ const addToLocalStorage = () => {
 
 const addToTasks = (e) => {
   e.preventDefault();
-  tasks.push({
+  let item = {
     description: taskDescription.value,
     completed: false,
     index: tasks.length,
-  });
+  }
+  tasks.push(item);
+  createUi(item);
   addToLocalStorage();
-  window.location.reload();
   taskDescription.value = '';
 };
 
@@ -49,8 +121,8 @@ const deleteTask = (listId) => {
 const deleteAllCompleted = () => {
   const uncompleted = [];
   for (let i = 0; i < tasks.length; i += 1) {
-    if (tasks[i].completed === false) {
-      uncompleted.push(tasks[i]);
+    if (task.completed === false) {
+      uncompleted.push(task);
     }
   }
   localStorage.setItem('lstore', JSON.stringify(uncompleted));
